[All] Added Role and Policy classes that implement SQLWrapper. PgPolicy and PgRole extends from those now
